[Refactor] #22 - client, kis dir 위치 이동
